Move tools into cmd/, move libraries into internal/
