+++ cmd/ntt/main.go
@@ -1,9 +1,10 @@
-package cmd
+package main
 
 import (
 	"os"
 
-	"github.com/nokia/ntt/ttcn3/syntax"
+	"ntt/internal/ttcn3/syntax"
+
 	"github.com/spf13/cobra"
 )
 
@@ -17,14 +18,14 @@ var (
 	Verbose = false
 )
 
-func Execute() {
+func init() {
+	rootCmd.PersistentFlags().BoolVarP(&Verbose, "verbose", "v", false, "verbose output")
+	rootCmd.AddCommand(listCmd)
+}
+
+func main() {
 	if err := rootCmd.Execute(); err != nil {
 		syntax.PrintError(os.Stderr, err)
 		os.Exit(1)
 	}
-}
-
-func init() {
-	rootCmd.PersistentFlags().BoolVarP(&Verbose, "verbose", "v", false, "verbose output")
-	rootCmd.AddCommand(listCmd)
 }
