+++ internal/ttcn3/syntax/errors.go
