+++ internal/ttcn3/syntax/nodes.go
