+++ internal/ttcn3/syntax/parser.go
