+++ internal/ttcn3/syntax/parser_test.go
