+++ internal/ttcn3/syntax/position_test.go
