+++ internal/ttcn3/syntax/scanner.go
