+++ internal/ttcn3/syntax/scanner_test.go
