+++ internal/ttcn3/syntax/serialize.go
