+++ internal/ttcn3/syntax/serialize_test.go
