+++ internal/ttcn3/syntax/syntax.go
