+++ internal/ttcn3/syntax/token.go
