--- main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/nokia/ntt/internal/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
